Eliminate warnings (#701)
* Externals in angle brackets.

* Commons from lib

* Update to preview versions of externals.

* Update fronter

* Switch to std::filesystem::path

* Unused variables

* Type conversion stuff.

* constness

* Missing headers

* Prefix increment, not postfix

* Check optional

* Remove a stray extern

* explicit

* Return an error case.

* Default value

* Enforce warnings.

* Disable some unfixable warnings.

* Update commons

* Formatting

* Fix a test

* Use fixed common

* Fix tests.

* Formatting

* Update to merged versions

* Make codefactor happier.

+++ src/hoi4_world/roles/stories_creator.cpp
@@ -1,9 +1,10 @@
 #include "src/hoi4_world/roles/stories_creator.h"
 
+#include <external/fmt/include/fmt/format.h>
+
 #include <ranges>
 #include <set>
 
-#include "external/fmt/include/fmt/format.h"
 #include "src/hoi4_world/roles/roles_importer.h"
 
 
@@ -72,12 +73,12 @@ std::optional<std::vector<std::pair<Tag, CombinationName>>> SortCombinations(
     const std::map<std::string, hoi4::Role>& roles)
 {
    std::ranges::sort(combinations, [roles](const auto& a, const auto& b) {
-      int a_score = 0;
+      float a_score = 0.0F;
       if (const auto role = roles.find(a.second); role != roles.end())
       {
          a_score = role->second.GetScore();
       }
-      int b_score = 0;
+      float b_score = 0.0F;
       if (const auto role = roles.find(b.second); role != roles.end())
       {
          b_score = role->second.GetScore();
